+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),  
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).  
   
+## [14.0.0] - 2023-10-06  
+### Changed  
+- `Shortcuts` class renamed to singular `Shortcut`.  
+- `yandexGames.sdk.features.LoadingAPI.ready` method is no longer called automatically to prevent game moderation rejection.  
+- `GameReady` class contents are moved to `YandexGamesSdk` class.  
+  
+### Added  
+- Added `Shortcut` and `GameReady` testing buttons to playtesting canvas.  
+  
 ## [13.3.0] - 2023-09-27  
 ### Added  
 - Added GameReady API that allows to manually call `ysdk.features.LoadingAPI.ready()` when the game has finished loading.  
@@ -119,6 +128,7 @@ This was leading to errors that were difficult to reproduce and detect. Now you 
 - Renamed RequestProfileDataPermission to RequestPersonalProfileDataPermission.  
 - Renamed HasProfileDataPermission to HasPersonalProfileDataPermission.  
   
+[14.0.0] https://github.com/forcepusher/com.agava.yandexgames/compare/13.3.0...14.0.0  
 [13.3.0] https://github.com/forcepusher/com.agava.yandexgames/compare/13.2.0...13.3.0  
 [13.2.0] https://github.com/forcepusher/com.agava.yandexgames/compare/13.1.0...13.2.0  
 [13.1.0] https://github.com/forcepusher/com.agava.yandexgames/compare/13.0.0...13.1.0